pyproject.toml
[project]
name = "palmoni"
version = "0.1.1"
description = ""
license = {text = "MIT"}
authors = [
    {name = "daniel-wallace-personal",email = "wallace.danielk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pynput (>=1.8.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "typer (>=0.16.1,<0.17.0)",
    "duckdb (>=1.3.2,<2.0.0)"
]

[project.scripts]
palmoni = "palmoni_core.cli.commands:main"

[tool.poetry]
packages = [{include = "palmoni_core"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"

[tool.pytest.ini_options]
testpaths = ["palmoni_core/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
